[#22] Split Grog.Client into Grog.Client and Grog.Client.Tasks
